drone/requirements.txt
flask
flask-cors
requests
psutil
opencv-python
inference-sdk
numpy
pymavlink
pyserial
stm32loader